requirements.txt
langchain
langchain-community
langchain-openai
langchain-astradb
astrapy
python-dotenv
django
pypdf
daphne
channels
djangorestframework
django-cors-headers
channels
django_extensions
transformers
torch
psycop2-binary
pgvector